backend/requirements.txt
Django==5.0.6
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.2
django-cors-headers==4.4.0
django-filter==24.3
django-environ==0.11.2

channels==4.1.0
channels-redis==4.2.0
daphne==4.1.2
uvicorn==0.30.1
gunicorn==22.0.0

celery==5.4.0
redis==5.0.7

psycopg[binary]==3.2.1
dj-database-url==2.2.0

ncclient==0.6.16
httpx==0.27.0
minio==7.2.7
weasyprint==62.3

structlog==24.2.0
django-prometheus==2.3.1

pytest==8.3.2
pytest-django==4.8.0
pytest-asyncio==0.23.8
factory-boy==3.3.0
ruff==0.5.5
